[deploy] Own Finelog command dispatch
